{Update} Stages & LevelGenerator
Stages: grass and dirt are now composed of 3*3 sprites (-> a list of 3 Sprite[3]). Each "bloc" type has to follow the same format.

LevelGenerator: When generating a tile, looks at it's up/down/left/right neighbors to see if it's the same type and shoose the right Sprite accordingly. (It's not yet perfect, but leagues better than without)